Tests/LibWeb/Text/input/css/serialize-shorthand-with-conflicting-logical-alias.html

<!DOCTYPE html>
<style>
    div {
        padding: 1px;
        padding-inline-start: 1px;
        padding-bottom: 1px;
    }
</style>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
    });
</script>
