Tests/LibWeb/Text/expected/css/serialize-shorthand-with-conflicting-logical-alias.txt

div { padding-top: 1px; padding-right: 1px; padding-left: 1px; padding-inline-start: 1px; padding-bottom: 1px; }
